+++ Pandora/Pattern/Functor/Covariant.hs
@@ -63,6 +63,3 @@ class Covariant (t :: * -> *) where
 	(<&&&&>) :: (Covariant u, Covariant v, Covariant w)
 		=> t :. u :. v :. w := a -> (a -> b) -> t :. u :. v :. w := b
 	x <&&&&> f = f <$$$$> x
-
-instance Covariant ((->) a) where
-	(<$>) = (.)