+++ Pandora/Pattern/Functor/Pointable.hs
@@ -1,12 +1,8 @@
 module Pandora.Pattern.Functor.Pointable where
 
 import Pandora.Core.Functor (type (|->))
-import Pandora.Core.Morphism ((!))
 import Pandora.Pattern.Functor.Covariant (Covariant)
 
 class Covariant t => Pointable t where
 	{-# MINIMAL point #-}
 	point :: a |-> t
-
-instance Pointable ((->) e) where
-	point = (!)